Ensure byte_addr_width is an int

+++ python/xfcp/node.py
@@ -186,13 +186,13 @@ class MemoryNode(Node):
             self.word_size = obj.word_size
             self.count_width = obj.count_width
 
-        self.byte_addr_width = self.addr_width+math.ceil(math.log(self.word_size/8, 2))
+        self.byte_addr_width = int(self.addr_width+math.ceil(math.log(self.word_size/8, 2)))
 
     def init(self, id_pkt=None):
         super(MemoryNode, self).init(id_pkt)
 
         self.addr_width, self.data_width, self.word_size, self.count_width = struct.unpack_from('<HHHH', self.id_pkt.payload, 2)
-        self.byte_addr_width = self.addr_width+math.ceil(math.log(self.word_size/8, 2))
+        self.byte_addr_width = int(self.addr_width+math.ceil(math.log(self.word_size/8, 2)))
 
         return self
